[API] Rudimentary EFK setup for development environment
Rudimentary setup in the Procfile and docker-compose.yml files to
support a basic installation of Elasticsearch, Kibana and Fluentd for
the development environment to encourage careful thought about logging
structure within the application.

Implementation captures output (using tee) of the individual
sub-programs and directs them to a directory of files which is available
to the Docker container running Fluentd as a moujnted volume.