[API] Clean-up env vars for config and ensure consistency of use.
This ensures that the values for the Git working directory, Operation
Log filesystem persistence dir, and for the "limits" cache path can be
set easily using the HAR_FILESYSTEM_OP_LOG_DIR
HAR_FILESYSTEM_GIT_TMP_DIR and HAR_LIMIT_STORE_CACHE_DIR variables
respectively.

It also fixes an issue where a path for the limit store cache path was
hard-coded in one function (persistence) but not hard coded in the
lookup function. In production this never mattered because the
hard-coded path happened to match the configured path in the previously
existing JSON config files.