Remove BrainTree proxy. Extract Limits to gCPC µService.
A mostly complete removal of the BrainTree *Subscription* dependency
with the important part of BrainTree "Customer" and "Credit Card" kept.

The "Limits" service, formerly tightly bound to the Braintree proxy (was
dependant on the proxy to query for an active subscription at BrainTree)
has been rewritten to be simpler, and care only for the "Billing Events"
at the Harrow database.

The Limits service, formerly a baked-in "service" that was in-process
has been extraced to be a gRPC mircoservice. gRPC has been chosen
because it avoids discussions about HTTP formatting, JSON serialization,
headers, authentication, authorization, etc and maps *quite* cleanly to
a simple language native RPC. The new introduced types (gRPC function
call types and service definitions) are somewhat less than ideal, but
not too intrusive.

The dependencies (Glide.lock) were also updated in this version. This
code is tested against Go 1.9