Add index decorator to handle missing values
Some operations in the activity stream refer to jobs for which there
are no activities.  This causes a problem when processing an
operation.created activity for such an operation:

    panic: Operation(3608c99c-fc0b-4799-4de7-30e3f53947e1).ProjectUuid(): 59faaa7d-582d-44fe-7b37-15482249a3c4:*projector.Job: not found [recovered]
            panic: Operation(3608c99c-fc0b-4799-4de7-30e3f53947e1).ProjectUuid(): 59faaa7d-582d-44fe-7b37-15482249a3c4:*projector.Job: not found

    goroutine 22 [running]:
    panic(0xd7ac00, 0xc823fb97f0)
            /usr/local/go/src/runtime/panic.go:481 +0x3e6
    bitbucket.org/harrowio/api/cmd/projector.(*BroadcastHandler).HandleActivity.func1(0xc820263300, 0xc82eba8380)
            /tmp/gopath/src/bitbucket.org/harrowio/api/cmd/projector/broadcast_handler.go:50 +0x1d7
    panic(0xd7ac00, 0xc823fb97f0)
            /usr/local/go/src/runtime/panic.go:443 +0x4e9
    bitbucket.org/harrowio/api/cmd/projector.(*Operation).ProjectUuid(0xc824b91c80, 0x7f3f400241e0, 0xc8202801c0, 0x0, 0x0)
            /tmp/gopath/src/bitbucket.org/harrowio/api/cmd/projector/operations.go:17 +0x279

The decorator IndexWithDefaults is introduced to fill in default
values when trying to retrieve an object with `IndexTransaction.Get`.

Since no concrete error classes are defined for index operations and
time constraints do not allow for introducing them, detecting that an
object could not be found is done via scanning the error message
returned by `Index.Get`.