Introduce improved operation runner.
This runner does not have a dependency on RabbitMQ as is uses PostgreSQL
9.5+ feature "SELECT ... SKIP LOCKED", which combined with "SELECT ..
FOR UPDATE" allows atomic locking within a transaction to ensure "at
most once" semantics.

The first-run behaviour has also been changed. The previous
operation-runner was not able to take work directly from the database,
relying instead on the message bus. Incase messages on the bus were
lost, the underlying database record would also be "lost" in the sense
that it would never be worked on. The new oepration runner solves this
problem by querying for the oldest unstarted operation in the datbabase,
checking it's TTL clauses, and either marking it as "TTL Expired" (no
more "cancelled by user" as in the old ver.) or running it.

The "running it" part remains the same, we "shell out" to the fsbuilder
and the controller-{lxd,shell} depending on whether it is a user job or
a notifier job. The shelling out has been cleaned up for the sake of
improved logging, and we now use proper Pipes from the IO package to
join the processes together which gives us finer-grained error
information incase there is a fault (e.g differentiating whether
fsbuilder or controller-{lxd,shell} had a problem), as well as the
ability to forward the "unused" std{out,err} to our log stream
(previously it was dropped).

An exponential back-off package has been introduced to help relieve load
on LXD, combined with improvements upstream, we should no longer "leak"
containers as before.

Enforcement of the maximum two hour running time is still guarded by
the controller-lxd, and the chance to remove the "clean-up" cron jobs on
the operation hosts means we should avoid the situation where a 10
minute waiting time leads to a container being killed after 1h50m and a
customer being denied their full run-time.