Ensure we don't start more than one operation
Race condition could allow us to take more work than we can handle.
Setting the channel to nil kills the goroutine because send on a nil
channel blocks indefinitely whilst send on a closed channel causes a
panic.

This logic probably needs another look, I believe in this sceheme we
leak goroutines which is probably harmless, but it's unclean.