Move "Mark As Started" and other logic to runOperation.
Referencing issue #40 this should address at least the issue with the
false starting. The operaion is only modified if we sucessfully proceed
to the runOperation goroutine.

Previously we could mark an operation as started, and then block forever
sending on a nil pendingOperations channel which has been closed because
we've raced into the for{ select {}} block.